+++ contracts/hub/router/src/testing/mod.rs
@@ -1,0 +1,3 @@
+pub mod fixtures;
+#[cfg(test)]
+pub mod helpers;
