--- contracts/hub/router/src/tests/mod.rs
@@ -1,2 +1,0 @@
-pub mod fixtures;
-pub mod tests;
